requirements.txt
flask
requests
python-dotenv
sentence-transformers
numpy
flask-login
flask-sqlalchemy